+++ src/components/GroupsEditor/GroupsEditor.test.tsx
@@ -2,7 +2,7 @@ import { toDataFrame } from '@grafana/data';
 import { act, fireEvent, render, screen, within } from '@testing-library/react';
 import { getJestSelectors } from '@volkovlabs/jest-selectors';
 import React from 'react';
-import { DragDropContext, DropResult } from 'react-beautiful-dnd';
+import { DragDropContext, DropResult } from '@hello-pangea/dnd';
 
 import { TEST_IDS } from '../../constants';
 import { LevelsEditor } from '../LevelsEditor';
@@ -36,10 +36,10 @@ jest.mock('../LevelsEditor', () => ({
 }));
 
 /**
- * Mock react-beautiful-dnd
+ * Mock @hello-pangea/dnd
  */
-jest.mock('react-beautiful-dnd', () => ({
-  ...jest.requireActual('react-beautiful-dnd'),
+jest.mock('@hello-pangea/dnd', () => ({
+  ...jest.requireActual('@hello-pangea/dnd'),
   DragDropContext: jest.fn(({ children }) => children),
   Droppable: jest.fn(({ children }) => children({})),
   Draggable: jest.fn(({ children }) =>