+++ src/components/LevelsEditor/LevelsEditor.tsx
@@ -9,7 +9,7 @@ import {
   Droppable,
   DropResult,
   NotDraggingStyle,
-} from 'react-beautiful-dnd';
+} from '@hello-pangea/dnd';
 
 import { TEST_IDS } from '../../constants';
 import { Level, LevelsGroup } from '../../types';